Deprecate 12-use-github-flow in favor of 24-do-scaled-trunk-based-development.
It has come to my attention that GitHub Flow is not a commercialized,
better known name for scaled trunk-based development, but in fact
a similar, slightly different workflow which dictates that deploys
must happen off feature branches before merging to master.
We do not want to dictate, at this time, at which step deploys should
take place, and are therefore replacing GitHub Flow with
scaled trunk-based-development.